fixed and refactored current account screen
